+++ resources/views/admin/subjects/list.blade.php
@@ -1,6 +1,6 @@
 @extends('admin.layout.index')
 @section('content')
-<!-- Page Content -->
+    <!-- Page Content -->
     <div class="container-fluid">
         <div class="row">
             <div class="col-lg-12">
@@ -15,28 +15,34 @@
                     <th>ID</th>
                     <th>Name</th>
                     <th>Edit</th>
-                    <th>Delete</th>
+                    @can('can-delete', 'user')
+                        <th>Delete</th>
+                    @endcan
                 </tr>
                 </thead>
                 <tbody>
                 @foreach($subjects as $subject)
-                <tr class="even gradeC" align="center">
-                    <td>{{$subject->id}}</td>
-                    <td>{{$subject->name}}</td>
-                    <td class="center">
-                        <button type="submit"><a href="{{route('subjects.edit', ['subject'=>$subject])}}">Edit</a></button>
-                    </td>
-                    <td class="center">
-                        {!! Form::open(['method'=>'DELETE', 'route'=>['subjects.destroy', 'subject'=>$subject]]) !!}
-                        <button type="submit" onclick="return confirm('Do you want to delete this field?')"><a>Delete</a></button>
-                        {!! Form::close() !!}
-                    </td>
-                </tr>
-                    @endforeach
+                    <tr class="even gradeC" align="center">
+                        <td>{{$subject->id}}</td>
+                        <td>{{$subject->name}}</td>
+                        <td class="center">
+                            <button type="submit"><a href="{{route('subjects.edit', ['subject'=>$subject])}}">Edit</a>
+                            </button>
+                        </td>
+                        @can('can-delete', 'user')
+                            <td class="center">
+                                {!! Form::open(['method'=>'DELETE', 'route'=>['subjects.destroy', 'subject'=>$subject]]) !!}
+                                <button type="submit" onclick="return confirm('Do you want to delete this field?')"><a>Delete</a>
+                                </button>
+                                {!! Form::close() !!}
+                            </td>
+                        @endcan
+                    </tr>
+                @endforeach
                 </tbody>
             </table>
         </div>
         <!-- /.row -->
     </div>
     <!-- /.container-fluid -->
-    @endsection
+@endsection
